fix: fold session gateway wiring into runtime_sessions for pkg-file-count
Merge session_gateway.go and session_gateway_transport_test.go into existing
pkg/service files so the package stays within the 15-file limit while
preserving session gateway host seams and BuildFactoryService wiring tests.
